ec/google/chromeec: Re-scope CREC under \_SB so that the Sync IRQ is accessible
Sync IRQ is not accessible under the limited resources of LPCB
in Windows.

IRQ is accessible under \_SB or \_SB.PCI0. However, since the EC
is not a PCIe device, it makes more sense under \_SB.

Move to \_SB so that sync IRQ doesn't cause a Code 12 in Windows

TEST=build+boot google/nocturne to Win11 and verify EC device shows
up and sync IRQ works for volume buttons in Windows

Change-Id: Ib7d6ef8214e72590851487f62faa96842814db3d
Signed-off-by: CoolStar <coolstarorganization@gmail.com>