ec/google/chromeec: Ensure pre-CR50 devices use "short" battery strings
Commit b97ec4f0166b ("chromeec: support reading long battery strings")
added support for reading long battery strings, but the method used to
determine if this feature is supported causes Windows to hang when
booting on older pre-CR50 ChromeOS devices. Add a short circuit to the
BRSX method to ensure that pre-CR50 devices use older/shorter strings.

In testing, affected platforms were Broadwell/Braswell and earlier.
Skylake is pre-CR50 as well, but unaffected. No adverse impact to
grouping it with the others as those boards do not support or use
longer battery strings.

TEST=build/boot Windows on google link, wolf, samus, chell and others.
Verify boot successful and battery info populated correctly.

Change-Id: If3686dc9b82b44202947081ec7e253317deb8fff
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>